+++ packages/odata-service-writer/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/odata-service-writer
 
+## 0.27.8
+
+### Patch Changes
+
+-   b9675bb: Improve Fiori Tools UI Texts
+
+## 0.27.7
+
+### Patch Changes
+
+-   Updated dependencies [f75b89d]
+    -   @sap-ux/project-access@1.30.4
+    -   @sap-ux/ui5-config@0.28.3
+    -   @sap-ux/mockserver-config-writer@0.9.4
+
 ## 0.27.6
 
 ### Patch Changes